Pusula client added,code fix

+++ bbt.gateway.common/Models/Database/PhoneConfiguration.cs
@@ -9,7 +9,7 @@ namespace bbt.gateway.common.Models
     {
         public Guid Id { get; set; } = Guid.NewGuid();
         public Phone Phone { get; set; } 
-        public int? CustomerNo { get; set; } 
+        public ulong? CustomerNo { get; set; } 
         public OperatorType? Operator { get; set; }
         public ICollection<PhoneConfigurationLog> Logs { get; set; }
         public ICollection<OtpRequestLog> OtpLogs {get;set;}
